Add labels and markers to the Street View panorama (#1081)
Show the same dot markers and fruit-name labels in the Street View
panorama as on the map, by extracting the map's marker/label rendering
into a shared module (locationMarkerHelpers.js) and reusing it. The
OverlayView labels attach to a StreetViewPanorama as well as a Map, so
both views render identically.

- All nearby locations get the map's dots (blue by default, orange when
  saved) plus their fruit-name label; labels use the satellite/hybrid
  style (white text, dark outline) since the panorama is imagery.
- The selected/viewed location additionally gets the orange "here" pin
  above its dot, using the same boxicons Map glyph as the map's MapPin.
- The map's own markers are hidden while Street View is open so they
  don't double up with the panorama's.
- Guard against async re-entrancy so a superseded panorama can't leave
  orphaned overlays on the shared panorama.

Closes #1081
